Add the debug leaks view to ParaView
The debug leaks view is be shown as a top level widget on application
startup if the environment variable PV_DEBUG_LEAKS_VIEW is set.

pqPythonDebugLeaksView extends the vtkQtDebugLeaksView to add Python
capabilities to the debugger.  You can add references to tracked VTK
objects to the Python console for inspection.

Change-Id: I26918d9a499817ff5c75cef5dba82e5ba1de9ae7