Execute python code in the context of the interactive console.
The function pqPythonShell::executeScript() is used to programmatically
execute python code.  This is used when running python scripts the user
selects with the Run Script button.  These should run in the console
context.

Code executed with vtkPythonInterpreter::RunSimpleString() will run in the
__main__ context which is now different from the ParaView python console.

The "from paraview.simple import *" statement is only executed in the
ParaView python console.

Change-Id: I5f957e6d60a6f77d37a6ed7c6b7cf1eec10fb517