Fix the SetParameter API overloads of the vtkPythonProgrammableFilter
Commit fc8cabcd (April 2007) introduced a SetParameter function on the
programmable filter.  It added a string vector property to filter proxy
that could be used to set arbitrary parameter name/value pairs that would
be defined in the filter python script.

Commit 8311d712 (July 2011) introduced several SetParameter overloads so
that custom string, int, and double vector properties were supported.
This allowed proxy xml to declare custom properties on the python
programmable filter that would be presented in the GUI like regular
properties.  But, this commit also changed the behavior of the original
Parameter string vector property.

Commit 9176ba5c82 (March 2012) accidentally broke the SetParameter
overloads.

This commit (May 2013) fixes the SetParameter overloads and original
Parameter string vector property (but both cannot be used together).
This commit adds a test for the SetParameter API and adds an example
proxy xml file that demonstrates how to define a python programmable
filter with custom parameters.

Change-Id: I0e6ff3e438f12305cae8b07dc7e61cb6988aebe0