Remove pqPythonShell::initPythonInerpreter callback.
Now, the pyPythonManager calls pqPythonShell::setupInterpreter() after
the pqPythonShell has been created.  This way, the pyPythonShell does
not execute an python code in its constructor, but the pqPythonManager
guarantees that it is fully initialized after construction.

Executing python code in the constructor is bad because of the way stdout
is handled in a callback by the pqPythonManager, and using a timer to
initialize later is bad because the stdout won't be routed correctly
until after the timer callback.

This commit also removes the ExitEvent handling from the pyPythonShell
since ParaView only finializes Python once at exit, there is no need to
have a handler to reset the pyPythonShell.

Change-Id: Ib9b53f38f4b7fc558823dcb11c029554bd23828c