fixed bug. currentdata display could point to wrong function or index in gaussianopacityfunctions.
index would not be set to -1 open deletion of a gaussian. this meant a gaussian with invalid index might be read.
Also, it was possible that a gaussian from the scalargaussianopacity could be selected after a gradient one was deleted.
When a gaussian was manipulated, the data field was not set to enabled.