fixed bug in setting gradient function ranges to data when connected to one or more servers.
WARNING: There is an odd, unkown bug than get the servers desynchronized in terms of what needs to be calculated.
This leads to servers waiting for an mpi message, when one will never be sent (one does calculations and waits,
the other never does any calculation and this everything is hunky dory). Currently the widgets manage to
circumvent this, but it really ought to be fixed.