fixed deadlock problem for vtkpvimageaccumulateinformation.
Removed the automatic update from the information gathering.
when copying from the object, the vtkpvimage... would update gradient range
and histogram if necessary. This caused a deadlock because the of the rootonly
being equals to 1. The root would try to update it's data, but couldn't get the
rest from the other nodes.